:sparkles: restructured GUI and added ability to plot nodes on a grid
